requirements.txt
Authlib==1.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
decorator==5.1.1
idna==3.4
pycparser==2.21
requests==2.28.2
tqdm==4.65.0
urllib3==1.26.15
validators==0.20.0
weaviate-client==3.23.2
